add Upstart 1.8 job file
:Release Notes:
This fix is simply moving the Upstart 1.8 job file for activitymanager
from the starfish-initscripts package to its proper home.

:Detailed Notes:
As mentioned above, this fix is just moving the upstart script - in the
initial implementation of Upstart 1.8, many 'component' upstart jobs
were modified for 1.8 and packaged with starfish-initscripts to ease
the transition from Upstart 0.3.11 to 1.8.  The plan was and it to
move the 'imported' scripts back to their home package as soon as is
reasonable.

:Testing Performed:
Since the file itself is functionally unchanged from that present in
starfish-initscripts, the testing done running Upstart 1.8 + s-i over
the last couple of months still applies.

The testing done specifically for this fix is to locally build the
component and inspect the output directory:
starfish/BUILD-goldfinger/work/armv7a-vfp-neon-starfish-linux-gnueabi/ \
   /activitymanager/2.0.0-110-r2/image/etc/init and ..../etc/event.d
to confirm that the properly configured upstart scripts end up in their
respective locations.

:QA Notes:

:Issues Addressed:
[GF-5620]  update coreOS upstart scripts for 1.8

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: I4d914fc4937d9f9cd387f840e63022749a6678c3
Reviewed-on: https://g2g.palm.com/2710
Reviewed-by: Build Verification
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Tested-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Roger Stringer <roger.stringer@lge.com>